refactor(test): making JSDoc more simple
